refactor(node)!: return an owned node and delete lifecycle handoff shims
Move startup, RPC composition, rollback, and ordered worker shutdown into
lifecycle owners. StartupGuard::finish returns the complete Node; delete
the detached-parts startup result, node_from_parts, cleanup forwarding,
and the run-module lifecycle paths without compatibility re-exports.

Delete the retired node-local DNS/outbound scheduler and unused P2P
listener-handle collection. P2pService remains the real networking owner.
Separate embedding progress and tests from the public embedding boundary.

Use a nonblocking teardown wake: a full channel already carries the wake.
Retain seven lifecycle regressions and the gateway-publication control;
add rollback, idempotence, queued-wake, and owned-startup-result coverage.
Update the embedding contract to the actual owner and teardown ordering.

Static source/whitespace checks only; compilation, rustfmt, Clippy, and
runtime tests remain unverified. No storage, consensus, dependency, or
workflow changes are included.

BREAKING CHANGE: internal startup returns Node, not detached parts; the
old run:: lifecycle paths and reconstruction/cleanup shims are removed.
