+++ crates/node/src/embed/testing.rs
@@ -1,0 +1,15 @@
+//! Polls synchronous node futures without adding an executor dependency.
+
+pub(crate) fn block_on<F: Future>(future: F) -> F::Output {
+    use std::task::{Context, Poll, Waker};
+
+    let mut future = std::pin::pin!(future);
+    let waker = Waker::noop();
+    let mut cx = Context::from_waker(waker);
+    loop {
+        match future.as_mut().poll(&mut cx) {
+            Poll::Ready(output) => return output,
+            Poll::Pending => std::thread::yield_now(),
+        }
+    }
+}
